refactor(login): component/container pattern
